#9 | Removed the THMApiManager
